:truck: rename: my ui 경로 수정
