+++ src/pages/my/ui/main/MyPage.tsx
@@ -2,8 +2,8 @@ import styled from '@emotion/styled';
 import { useState } from 'react';
 import Navigation from '@shared/components/Navigation.tsx';
 import Header from '@shared/components/Header.tsx';
-import ProfileSummary from '@widget/my/ui/ProfileSummary.tsx';
-import Menu from '@widget/my/ui/Menu.tsx';
+import ProfileSummary from '@widget/my/ui/main/ProfileSummary.tsx';
+import Menu from '@widget/my/ui/main/Menu.tsx';
 import type { HeaderIcon } from '@shared/types/header.ts';
 import type { MenuItemProps } from '@widget/my/types/menu.ts';
 import IcSet from '@icon/ic-set.svg';
