+++ src/widget/my/ui/main/Menu.tsx
